### 2023-03-31 #5171 @mstimberg
Hi @VincentStimper, thanks for the suggestions, I will start inviting reviewers later today.

> I would also volunteer to add myself to the list to potentially review for the JOSS in the future. Where can I do this?

Glad to hear that :blush: Please register here: https://reviewers.joss.theoj.org/join
