### 2023-04-12 #5361 @editorialbot
[REVIEW]: normflows: A PyTorch Package for Normalizing Flows
**Submitting author:** @VincentStimper (<a href="http://orcid.org/0000-0002-4965-4297">Vincent Stimper</a>)
**Repository:** https://github.com/VincentStimper/normalizing-flows
**Branch with paper.md** (empty if default branch):
**Version:** v1.7.0
