### 2023-04-24 #5361 @mstimberg
:wave: @kazewong, @matejgrcic no rush for performing the actual review of course, but just checking in to see whether there are any questions on your side or anything else I can do to help you with the review? Please be reminded that JOSS reviews can be iterative, i.e. feel free to open issues or comment here as soon as you come across concerns – there is no need to gather everything into a single review text as commonly done for traditional journals. In the same vein, please feel free to tick boxes in your checklist whenever you are sure about them (the two in the beginning – conflict of interest and code of conduct – would be good first candidates :wink: ). Thanks for your time :pray: !
