### 2023-04-27 b0d2237
Fixed sign of log det of forward pass
