### 2023-04-27 8b5415f
Fixed sign of log det of forward pass (#28)
