+++ setup.py
@@ -2,7 +2,7 @@ from setuptools import setup, find_packages
 from codecs import open
 from os import path
 
-__version__ = "1.6.1"
+__version__ = "1.6.2"
 
 here = path.abspath(path.dirname(__file__))
 
