### 2023-05-12 #5361 @mstimberg
Hi it's me again :blush:  @kazewong, @matejgrcic, how are things going from your side, anything I can do to help? Please note that at JOSS we strive to complete a first round of reviews within 6 weeks, and this review issue has been opened about 4 weeks ago. Could you give me some estimate when you think will be able to perform the review? Thanks a lot, and as always, don't hesitate to ask in case anything about the process is unclear :pray:

PS: I will be travelling for the next two weeks, so my replies might take a bit longer than usual. But as I said earlier, reviews here can be iterative, and authors and reviewers are free to discuss and fix issues without my direct involvement :wink:
