### 2023-05-30 VincentStimper/normalizing-flows#30 @kazewong
No community contribution guideline
Raising issue as part of JOSS review https://github.com/openjournals/joss-reviews/issues/5361

Currently, there is no community contribution guideline I can find. Per Joss review checklist:

There should be a standard template, it would be good to have an explicit contribution guide in the root directory
