### 2023-05-31 VincentStimper/normalizing-flows#31 @kazewong
Comments on JOSS paper
Raising issue as part of JOSS review https://github.com/openjournals/joss-reviews/issues/5361

I will put comments related to the JOSS paper itself in this thread. Overall the draft is concise and clear about the package and its statement of need, though there are a couple of things I would like the authors to clarify ( I am putting them in on a rolling basis, still have some deadlines to catch!) :

1. In the statement of need, the authors claim other packages such as distrax do not support autoregressive flow. While not having explicitly saying they offer an autoregressive flow, distrax actually has some lower/upper triangular matrix bijector (https://github.com/deepmind/distrax/tree/master/distrax/_src/bijectors), which I believe corresponds to an autoregressive flow. Considering the design choice of distrax are just the bijectors and distribution in jax, I would not phrase it such features do not exist in their code, even though it is not extensively marketed. I suggest rewording the first part of the statement of need to soften the tone .
