### 2023-05-31 7584bba
[Issue #31] Changing formulation in paper to be less ambiguous.
