### 2023-05-31 2eb6cb2
[Issue #30] Added guideline on getting support and contributing to the README
