### 2023-05-31 VincentStimper/normalizing-flows#30 @VincentStimper
Hi @kazewong,

given that this is a GitHub repository, I thought it would be self-explanatory that you can get help by submitting an issue and contribute via PRs. However, I agree that a guideline will help in any case.
Hence, [I added](https://github.com/VincentStimper/normalizing-flows/commit/2eb6cb2b1a46835876e219363d685316442acf2b) a guideline for getting support and contributing to the package to the [README](https://github.com/VincentStimper/normalizing-flows/blob/master/README.md).

Please let me know whether this is sufficient.

Best regards,
Vincent
