### 2023-05-31 VincentStimper/normalizing-flows#32 @matejgrcic
Example usage for images
Raising issue as part of JOSS review https://github.com/openjournals/joss-reviews/issues/5361

Please consider adding example usage (Colab notebook) for image data e.g. CIFAR10 or MNIST.
