### 2023-05-31 VincentStimper/normalizing-flows#33 @matejgrcic
More functionality
Raising issue as part of JOSS review https://github.com/openjournals/joss-reviews/issues/5361

Could you please show how could one implement multi-scale architecture for normalizing flows?
If it is not possible, please add this functionality.

Please consider adding augmented normalizing flows (https://arxiv.org/pdf/2002.09741.pdf, https://arxiv.org/pdf/2002.07101.pdf)
