### 2023-05-31 c41ad54
[Issues #32 and #33] Added Colab notebook to demonstrate multi-scale architecture applied to imaged
