### 2023-05-31 5f1906e
[Issue #33] Added target for augmented normalizing flow and example notebook of how to use it
