### 2023-05-31 VincentStimper/normalizing-flows#33 @VincentStimper
Hi @matejgrcic,

the multi-scale architecture is already implemented in this packages as the [`nf.MultiscaleFlow`](https://github.com/VincentStimper/normalizing-flows/blob/705edccf60853ffbd4ab8198df6412f65fb9f396/normflows/core.py#L300). Usage is demonstrated in the [image generation example notebook](https://github.com/VincentStimper/normalizing-flows/blob/master/examples/glow.ipynb) as well as in the corresponding [Colab notebook](https://github.com/VincentStimper/normalizing-flows/blob/master/examples/glow_colab.ipynb) I created for [this issue](https://github.com/VincentStimper/normalizing-flows/issues/32).

I just added a [target distribution](https://github.com/VincentStimper/normalizing-flows/blob/705edccf60853ffbd4ab8198df6412f65fb9f396/normflows/distributions/target.py#L76) that can be used to train an augmented normalizing flow, among other things. I also added a [notebook](https://github.com/VincentStimper/normalizing-flows/blob/master/examples/augmented_flow.ipynb) showcasing this.

Best regards,
Vincent
