### 2023-05-31 #5361 @matejgrcic
Hi, I have finished my review. The author addressed all of my concerns.
I believe that this package is a great contribution to future research on normalizing flows.

Best,
Matej
