### 2023-05-31 VincentStimper/normalizing-flows#35 @kazewong
Putting examples in the documentation 
Raising issue as part of JOSS review https://github.com/openjournals/joss-reviews/issues/5361

The examples in the repo are currently hosted under the https://github.com/VincentStimper/normalizing-flows/tree/master/examples folder. I think it would be nice to add a page in the documentation site that display these ipython notebook https://vincentstimper.github.io/normalizing-flows/, just like the API page.

Note that this issue is a nice to have that I don't think is necessary for JOSS review, so it shouldn't block the review process.
