### 2023-06-01 VincentStimper/normalizing-flows#35 @VincentStimper
Hi @kazewong,

good idea, I'm just wondering how to realize this. We could just add a table to the README with links to all the notebooks.
Other than that, is there a way to show the example notebooks in the documentation, @vincentberenz?

Best regards,
Vincent
