### 2023-06-02 ef73663
[Issue #36] Added forward and inverse method with log determinant to MultiscaleFlow
