### 2023-06-05 VincentStimper/normalizing-flows#35 @vincentberenz
@VincentStimper
There are two ways of doing this.

## mkdocs-jupyter
See: https://github.com/danielfrg/mkdocs-jupyter
This allows to add a visual html representation of the notebook in the documentation.
pros: We just need to setup the configuration of mkdocs once, and then things will be updated automatically every time a notebook is modified (or added, or deleted)
cons: This looks nice, but no code can be executed.
effort: I will need to find out how to make a nice layout in the menu, should not be too hard

## jupyter-cloud.gwdg.de
See: https://jupyter-cloud.gwdg.de
This is a server on which MPI employees can upload (and make public) notebook.
So we can upload the notebooks there, and provide a page in the documentation linking to them
pros: Code can be executed
cons: We need to reupload manually the notebooks everytime we update them
effort: A bit unclear how to make things public, and if dependencies package can be installed for each user.

Let me know which solution you would prefer
