### 2023-06-05 VincentStimper/normalizing-flows#35 @VincentStimper
Hi @vincentberenz,

I don't think the notebooks have to be executable. For this purpose, we already have the Google Colab notebooks. Hence, the first solution sounds just right to me.

Can we leave the notebooks in the current folder outside the package folder? It would be great if you had a look at how to make a nice layout.

Best regards,
Vincent
