### 2023-06-06 1142fb4
added jupyter notebooks to the html documentation
