### 2023-06-06 d43593e
replacing in docs symbolic links to file by symbolic link to folder
