+++ mkdocs.yml
@@ -16,15 +16,15 @@ nav:
   - About: index.md
   - API: references.md
   - Examples:
-    - augmented_flow: notebooks/augmented_flow.ipynb
-    - change_base_distribution: notebooks/change_base_distribution.ipynb
-    - circular_nsf: notebooks/circular_nsf.ipynb
-    - comparison_plan_rad_aff: notebooks/comparison_plan_rad_aff.ipynb
-    - glow: notebooks/glow.ipynb
-    - image: notebooks/image.ipynb
-    - neural_spline_flow: notebooks/neural_spline_flow.ipynb
-    - paper_example_nsf: notebooks/paper_example_nsf.ipynb
-    - planar: notebooks/planar.ipynb
-    - real_nvp: notebooks/real_nvp.ipynb
-    - residual: notebooks/residual.ipynb
-    - vae: notebooks/vae.ipynb
+    - augmented_flow: examples/augmented_flow.ipynb
+    - change_base_distribution: examples/change_base_distribution.ipynb
+    - circular_nsf: examples/circular_nsf.ipynb
+    - comparison_plan_rad_aff: examples/comparison_plan_rad_aff.ipynb
+    - glow: examples/glow.ipynb
+    - image: examples/image.ipynb
+    - neural_spline_flow: examples/neural_spline_flow.ipynb
+    - paper_example_nsf: examples/paper_example_nsf.ipynb
+    - planar: examples/planar.ipynb
+    - real_nvp: examples/real_nvp.ipynb
+    - residual: examples/residual.ipynb
+    - vae: examples/vae.ipynb
