### 2023-06-06 29a480c
added jupyter notebooks to the html documentation (#38)
* added jupyter notebooks to the html documentation

* the mkdocs workflow installs mkdocs-jupyter

* replacing in docs symbolic links to file by symbolic link to folder

---------

Co-authored-by: Vincent Berenz <vincent.berenz@tuebingen.mpg.de>
