+++ .github/workflows/mkdocs.yaml
@@ -13,5 +13,5 @@ jobs:
       - uses: actions/setup-python@v2
         with:
           python-version: 3.x
-      - run: pip install mkdocs-material mkdocstrings[python]
+      - run: pip install mkdocs-material mkdocstrings[python] mkdocs-jupyter
       - run: mkdocs gh-deploy --force
