### 2023-06-06 VincentStimper/normalizing-flows#35 @VincentStimper
Hi @kazewong,

@vincentberenz has added this feature, thanks a lot!
You're welcome to have a look at the [current documentation](https://vincentstimper.github.io/normalizing-flows/) yourself and close the issue if this is what you had in mind.

Best regards,
Vincent
