### 2023-06-07 #5361 @mstimberg
Hi again @VincentStimper. Everything so far looks in good shape to me, but could you consider to:

1. Add ORCIDs to the other authors as well (not mandatory but recommended)
2. Add DOIs to the Noé et al. reference, as well as to the arxiv preprints (which now [all have DOIs](https://blog.arxiv.org/2022/02/17/new-arxiv-articles-are-now-automatically-assigned-dois/))

Thanks :pray:
