+++ paper/paper.bib
@@ -61,7 +61,8 @@
   volume={365},
   number={6457},
   year={2019},
-  publisher={American Association for the Advancement of Science}
+  publisher={American Association for the Advancement of Science},
+  doi={10.1126/science.aaw1147}
 }
 
 @article{kobyzev2021,
@@ -152,7 +153,8 @@
     title = {Tensor{F}low {D}istributions},
     author = {Joshua V. Dillon and Ian Langmore and Dustin Tran and Eugene Brevdo and Srinivas Vasudevan and Dave Moore and Brian Patton and Alex Alemi and Matt Hoffman and Rif A. Saurous},
     journal = {arXiv preprint arXiv:1711.10604},
-    year = {2017}
+    year = {2017},
+    doi={10.48550/arXiv.1711.10604}
 }
 
 @software{deepmind2020jax,
@@ -211,11 +213,11 @@
   year      = {2020}
 }
 
-@article{Midgley2022,
+@article{Midgley2023,
   title={{Flow Annealed Importance Sampling Bootstrap}},
   author={Laurence Illing Midgley and Vincent Stimper and Gregor N. C. Simm and Bernhard Schölkopf and José Miguel Hernández-Lobato},
-  journal={arXiv preprint arXiv:2208.01893},
-  year={2022}
+  journal={International Conference on Learning Representations},
+  year={2023}
 }
 
 @article{boltzgen,
