### 2023-06-12 #5361 @VincentStimper
Hi @mstimberg,

I added the DOIs and ORCIDs.

Is there anything else I should add or change? If not, I would create a new release of the package to be referenced in the published article.

Best regards,
Vincent
