### 2023-06-12 #5361 @mstimberg
> Is there anything else I should add or change? If not, I would create a new release of the package to be referenced in the published article.

If you could also add the DOI for the Tabak&Turner paper as suggested by editorialbot above that would be great. Otherwise all looks good from my side, feel free to do the release and archive it (e.g. on Zenodo), please.
