+++ paper/paper.bib
@@ -18,7 +18,8 @@
   number={2},
   pages={145--164},
   year={2013},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/cpa.21423}
 }
 
 @inproceedings{rezende2015variational,
