+++ normflows/__init__.py
@@ -10,4 +10,4 @@ from . import utils
 from . import sampling
 from .sampling import HAIS
 
-__version__ = "1.6.2"
+__version__ = "1.7.0"
