### 2023-06-12 #5361 @mstimberg
Hi @VincentStimper, thanks for the changes and the release, we are almost there! I noticed that the Zenodo record lists the licence "Other" instead of MIT, could you change this manually in the archive metadata (no need for a new release)?
