### 2023-06-12 #5361 @mstimberg
Thanks for the changes. Apologies for the nitpicking, but it seems that the Zenodo archive incorrectly associates Bernhard Schölkopf's ORCID with José Miguel Hernández-Lobato. As soon as this is fixed we will be ready to go.
